pyproject.toml
[project]
name = "testprj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agentops>=0.4.16",
    "chainlit>=2.5.5",
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
]

[project.scripts]
testprj = "testprj:main"
llm= "testprj.simpleAgent:runAgent"
tool = "testprj.toolCalling:testAsync"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
